backend/requirements.txt
fastapi
APScheduler
Pillow
psycopg2-binary
python-dotenv
python-jose
python-multipart
uvicorn
pytz              
bcrypt                         
email-validator                                         
passlib                          
pip
websockets                                           
schedule
python-dateutil
requests
pydrive2
shapely
lxml
pydantic
